Add a <div> in the content section
Add a div tag in the content section of the html file with a distinct size and background color.

This serves as the laywork to get a grid of divs that will do the sketch. By now, just having a simple div that exists is enough.

The div is created by the DOM methods of java script.